requirements.txt
requests
python-dotenv
pandas
psycopg2-binary
python-dateutil
scikit-learn
joblib
textblob
faker
